Made some minor bug fixes and some neat features
Fixed some if statements to not throw the username in use before it was actually in use and a similar if statement in the signup page. Another change are thebuttos for the catalog page. I changed how the buttons looked by rounding the edges, giving them a dimgray trim, and lastly making them dim when you mouse over the buttons to show which one you're above. I made a change in the additem.php that allows me to avoid redundant entries into the database for the orders. And finally in this commit I made the cart css responsive more so for the area that shows things in the cart because it would cause an overflow on mobile devices.